Move helpers only used in tests to conftest.py
