+++ tests/test_stable/test_assets.py
@@ -2,8 +2,8 @@ import pandas as pd
 import pytest
 
 from cognite import CogniteClient
-from cognite.client._utils import generate_random_string
 from cognite.client.stable.assets import Asset, AssetListResponse, AssetResponse
+from tests.conftest import generate_random_string
 
 assets = CogniteClient().assets
